requirements.txt
aiofiles==22.1.0
asgiref==3.6.0
colorama==0.4.6
dict-toolbox==2.3.1
dj-database-url==1.2.0
Django==4.1.4
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
djangorestframework==3.14.0
gunicorn==20.1.0
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.1
msgpack==1.0.4
nest-asyncio==1.5.6
pop==23.0.0
pop-config==10.2.3
pop-loop==1.0.6
psycopg2-binary==2.9.5
pytz==2022.7
PyYAML==6.0
rend==6.5.1
sniffio==1.3.0
sqlparse==0.4.3
toml==0.10.2
whitenoise==6.3.0
